Implemented unit test for AddItem.
"on-chain" integration tests (with solana-program-test) are very
complicated and hard to debug.

I want to move as much as possible to unit tests, i.e.
solana-program-test will be used only for runtime-checked invariants
(for instance, writing to account, not owned by the program) and already
implemented correct inputs.

With this approach we will have easy to debug unittests and integration
test for correct inputs.